# Conflicts:
#	app/src/main/java/com/hll_sc_app/app/SplashActivity.java

+++ app/src/main/java/com/hll_sc_app/bean/report/lack/CustomerLackBean.java
@@ -4,7 +4,7 @@ public class CustomerLackBean {
 
     private int deliveryLackKindNum;
     private double deliveryLackAmount;
-    private int deliveryLackNum;
+    private double deliveryLackNum;
     private String purchaserID;
     private String purchaserName;
     private String shopID;
@@ -26,11 +26,11 @@ public class CustomerLackBean {
         this.deliveryLackAmount = deliveryLackAmount;
     }
 
-    public int getDeliveryLackNum() {
+    public double getDeliveryLackNum() {
         return deliveryLackNum;
     }
 
-    public void setDeliveryLackNum(int deliveryLackNum) {
+    public void setDeliveryLackNum(double deliveryLackNum) {
         this.deliveryLackNum = deliveryLackNum;
     }
